+++ src/game/Item.java
@@ -2,7 +2,9 @@ package game;
 
 public class Item {
 	private String name;
-	private float price;
+	private double price;
+	
+
 	
 	/* Getter and Setters for
 	 * name
@@ -18,11 +20,11 @@ public class Item {
 		this.name = name;
 	}
 
-	public float getPrice() {
+	public double getPrice() {
 		return price;
 	}
 
-	public void setPrice(float price) {
+	public void setPrice(double price) {
 		this.price = price;
 	}
